+++ app/routes.php
@@ -98,6 +98,7 @@ Route::group(['prefix' => config('thinktomorrow.chief.route.prefix', 'admin'), '
      */
     Route::post('managers/{key}/{id}/media', 'Thinktomorrow\Chief\App\Http\Controllers\Back\Media\UploadManagersMediaController@store')->name('chief.back.managers.media.upload');
     Route::get('api/internal-links', 'Thinktomorrow\Chief\App\Http\Controllers\Api\InternalLinksController@index')->name('chief.api.internal-links');
+    Route::get('api/media', 'Thinktomorrow\Chief\App\Http\Controllers\Api\MediaGalleryController@index')->name('chief.api.media');
 
     /**
      * -----------------------------------------------------------------