+++ resources/views/back/_elements/slimcropper-component.blade.php
@@ -7,15 +7,16 @@
     <script src="{{ asset('/assets/back/js/vendor/slim.js') }}"></script>
     <script>
         Vue.component('slim', {
-            props: ['options', 'group'],
+            props: ['options', 'name', 'group'],
             template: `
                     <div class="thumb">
                         <div class="slim">
                             <img v-if="url" :src="url" :alt="filename">
-                            <input v-if="id" type="file" :name="group+'[replace]['+id+']'" accept="image/jpeg, image/png, image/bmp, image/svg+xml, image/webp, image/gif"/>
-                            <input v-else type="file" :name="group+'[new][]'" accept="image/jpeg, image/png, image/bmp, image/svg+xml, image/webp, image/gif" />
+                            <input style="margin-bottom:0" v-if="id && !newUpload" type="file" :name="name+'[replace]['+id+']'" accept="image/jpeg, image/png, image/bmp, image/svg+xml, image/webp, image/gif"/>
+                            <input style="margin-bottom:0" v-if="newUpload" type="hidden" :name="name+'[new]['+ id +']'" :value="id" accept="image/jpeg, image/png, image/bmp, image/svg+xml, image/webp, image/gif" />
+                            <input style="margin-bottom:0" v-else type="file" :name="name+'[new][]'" accept="image/jpeg, image/png, image/bmp, image/svg+xml, image/webp, image/gif" />
                         </div>
-                        <input v-if="deletion" type="hidden" :name="group+'[delete][]'" :value="id"/>
+                        <input v-if="deletion" type="hidden" :name="name+'[delete][]'" :value="id"/>
                     </div>
                 `,
             data: function () {
@@ -26,29 +27,34 @@
                     filename: this.options.filename || null,
                     instance: null,
                     deletion: false,
+                    newUpload: false
                 }
             },
             mounted: function () {
-
                 this.options.didRemove = this.markForDeletion;
                 this.options.didLoad = this.onLoad;
+                this.newUpload = this.options.newUpload;
                 this.instance = new Slim(this.$el.childNodes[0], this.options);
 
+
                 // If a file instance is passed, we want to directly load the file into our cropper
-                if (this.file) this.instance.load(this.file);
+                if (this.file) {
+                    this.instance.load(this.file);
+                }
             },
             methods: {
-                markForDeletion: function (e) {
-                    // If event contains a reference to an image,
-                    // this is a replace action and not a delete one
+                markForDeletion: function (e, target) {
                     this.deletion = true;
+                    var self = this;
+
+                    Eventbus.$emit('file-deletion-' + this.group, {id: self.id, newImage: target});
                 },
                 onLoad: function () {
 
                     // Unmark for deletion
                     this.deletion = false;
 
-                    Eventbus.$emit('files-loaded-' + this.group,{});
+                    Eventbus.$emit('files-loaded-' + this.group, this.id);
 
                     // Let Slim know it's good to go on - didLoad callback allows for input check prior to Slim.
                     return true;