+++ src/Media/DuplicateAssetException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Thinktomorrow\Chief\Media;
+
+class DuplicateAssetException extends \Exception
+{
+}